confirm: only pause specific confirmations
Pausing a confirmation currently pauses all confirmations, not just
other instances of the current confirmation. For example, when trading,
pausing a Mark All confirmation will also skip confirmation of the Seize
action.

The prompt in showYesNoPrompt is "Pause this confirmation". To better
match that description, only pause new occurrences of the current
confirmation.