confirm: use interface rect for order-remove calculations
When the DF interface percentage is not 100, the interface width can be
smaller than the window width. For certain sizes (depending on the
interface percentage), the following condition can hold:

- interface width <= 154 < window width

In this situation, the info window tab row will not have "unwrapped"
(from four to two UI rows), but the previous order index calculation
code would assume that it had (due to using the window width). This two
UI row discrepancy caused the calculated order index to be one too high
(and possibly out of bounds) when clicking on either of the bottom two
UI rows of an order remove button. The incorrect index caused the
confirmation to display the description of the following order.